add a new version of import_ecm_results
This change provides two versions of the `import_ecm_results` function,
the older version could read in the results for a full ecm_results.json
file in about 12 seconds, the new version in about 2 seconds.
This commit is here to save both versions with the timing difference and
ability to check for differneces detween the two.  The next commit will
remove the old version leaving only the faster version of the method.

./Results_Files_3/ecm_results_1-1.json imported and coerced to a DataFrame in 0:00:00.927231 # Old
./Results_Files_3/ecm_results_1-1.json imported and coerced to a DataFrame in 0:00:00.360174 # New

./Results_Files_3/ecm_results_2.json imported and coerced to a DataFrame in 0:00:13.038668   # Old
./Results_Files_3/ecm_results_2.json imported and coerced to a DataFrame in 0:00:02.759039   # New

./Results_Files_3/ecm_results_3-1.json imported and coerced to a DataFrame in 0:00:13.225869 # Old
./Results_Files_3/ecm_results_3-1.json imported and coerced to a DataFrame in 0:00:02.867780 # New